Add line number to rules and strings.
With this change rules and strings have a field that indicates the line number where the rule or string was declared in the source code. This commit also makes the code a bit more clear by making the scanner to return the values associated to each token together with the token itself, instead of using the scanner's context for that.