Update .gitignore and cleanup Makefile (#39)
Update the .gitignore files to include *.swp and j2y.

While here, also fixup the Makefile as the release target is broken. It depends
upon two targets (parser and lexer) that do not exist. I'm guessing this
Makefile is not really used and is a holdover from the Northern-Lights repo?