Check for duplicate string identifiers.
Fix an issue where duplicate string identifiers were not properly causing an
error when parsing the rules.

In the process of adding this I discovered a test I had written had in a
previous commit had duplicate string identifiers. I also discovered the test for
the base64 alphabet length was slightly incorrect in how it was doing
assertions.