Add setErrorWithLineNumber().
Move the check for duplicate string identifiers into the rule creation and out
of the string creation part of the grammar. This way we can tell you which
string it is, the line number and the rule to which that string belongs.

While here, fix a misnamed rule in a test case from previous commit.